Refactor: architectural alignment and UI specs
Simplified layers by moving state to Domain, applied Figma colors, and modularized UI components to eliminate spaghetti code.
